fix(detection): no more problem with python paths
It  was occuring some errors when importing the detect.py file and
other personal modules and packages like `utils`due to the
sys.path.appen that was being done.

This commit fix it by bringing yolov5 folder inside this project
repository and calling it directly without forcing to the PATH.

OBSERVATION:
    The ideia behind having a specific `detect.py` file at the top level of this
project is to have an easy and fast way to make detections inside this
project by calling it directly with changes inside it.

ALSO:
    It is added `yolov5` folder to `.gitignore`