feat: add a video parser to extract information about relevant events
The old interaction_detector has been renamed to model_inference.
A video parser is added. And things were restructured.

The `model_inference` contains code related to the creation of the
detection model, results of the inference + treatment and materialization
of these results in bounding boxes.

The `video_parser` is responsible to generate the bboxes for each frame,
metrics for actors proximity and point which frames contain interaction
between the actors by looking at the bboxes generated. It also generate
statistics about the video that are important to store as patient's
evolution