feat(app): add CRUD to streamlit interactive page
Now the user can add the session(composed of toddler, video and date),
to the database.

The videos are stored in ".../database/videos/" only if the user decides
to add the session to the DB.

The strealit page does not interact with the DAOs. This is the work of
the "st_controll.py" file. It is just a quick way to separete the page
from the database, not the best way, but helps in readabillity.