refactor: remove references to local dev system
To avoid references to absolute paths from our local system in production
code (when the app is up and running in the cloud), those references
were modified.

For the YOLO path: With Torch Hub, we can pass a github repo, and not
just a local repo. Thus it is passed the Ultrlytics' YOLOv5 github repo
on the v6.22version (that's the latest version we tested)

For the NN weights: relative paths are passed