fix(pages): fix creation of DB and squema
For some reason, when the database was being  deleted and the first
page to open was not the welcome entrypoint (that was supposed to create
the DB and squema) an empty database with no squema (??????????)
was being created.

I tried to dinf what could be creating this file, but saddly didn't succed.

What I did to avoid this was to try to create the database in the beginning of each
page, making sure to create a a Database AND a squema together, and not
some weird empty file (T﹏T)